% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhnmsd.R
\name{yhnmsd}
\alias{yhnmsd}
\title{yhnmsd()}
\usage{
yhnmsd(data, mydv, est_d = 2, ci_d = 2, propcut)
}
\arguments{
\item{data}{dataset}

\item{mydv}{response variable}

\item{est_d}{number of digits for estimate}

\item{ci_d}{number of digits for CI}

\item{propcut}{cutpoint to create proportion}
}
\value{
data.frame
}
\description{
yhnmsd()
}
