% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhnmsd.R
\name{yhnmsd}
\alias{yhnmsd}
\title{yhnmsd generates desriptive stats}
\usage{
yhnmsd(data, mydv, est_d = 2, ci_d = 2, propcut)
}
\arguments{
\item{data}{a dataframe}

\item{mydv}{response var}

\item{est_d}{est}

\item{ci_d}{ci}

\item{propcut}{cutpoint to generate prop}
}
\value{
a dataframe
}
\description{
yhnmsd generates desriptive stats
}
